requirements.txt
django==3.2
psycopg2-binary==2.8.6
Pillow==8.1
djangorestframework==3.11
django-rest-passwordreset==1.2.1
celery==5.2.3
drf-yasg==1.20.0
redis==4.1.1
python-decouple==3.5



